# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ClawdBot Marketplace API",
    "version": "1.0.0",
    "description": "Agent Discovery API for autonomous AI agents. A2A, Claude Marketplace, MCP, and OpenClaw protocols. 617+ tools, 14 on ClawHub.",
    "contact": {
      "name": "ClawdBot Support",
      "url": "https://clawdbot.com",
      "email": "support@clawdbot.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://clawdbot.com",
      "description": "Production server"
    },
    {
      "url": "https://gicm.app",
      "description": "Alternative production server"
    }
  ],
  "tags": [
    {
      "name": "Discovery",
      "description": "Agent discovery and marketplace catalog"
    },
    {
      "name": "Search",
      "description": "Search and filter marketplace items"
    },
    {
      "name": "Installation",
      "description": "Get installation instructions for items"
    },
    {
      "name": "A2A",
      "description": "Agent-to-Agent task submission and execution"
    },
    {
      "name": "Agents",
      "description": "External agent registration and discovery"
    },
    {
      "name": "Health",
      "description": "Service health and status monitoring"
    }
  ],
  "paths": {
    "/.well-known/claude-marketplace.json": {
      "get": {
        "tags": ["Discovery"],
        "summary": "Get full marketplace catalog",
        "description": "Returns complete registry of 617+ security-verified items with agent discovery metadata",
        "operationId": "getMarketplaceCatalog",
        "responses": {
          "200": {
            "description": "Successful response with full catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceCatalog"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search marketplace items",
        "description": "Search and filter items by query, kind, and tags",
        "operationId": "searchItems",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string",
            "schema": {
              "type": "string",
              "example": "solana"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "Filter by item kind",
            "schema": {
              "type": "string",
              "enum": ["agent", "skill", "mcp", "workflow", "command"],
              "example": "agent"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "description": "Filter by tag",
            "schema": {
              "type": "string",
              "example": "blockchain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketplaceItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - query parameter required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/install": {
      "post": {
        "tags": ["Installation"],
        "summary": "Get installation instructions",
        "description": "Returns installation instructions and security info for a specific item",
        "operationId": "getInstallInstructions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["item"],
                "properties": {
                  "item": {
                    "type": "string",
                    "description": "Item identifier in format 'category/slug'",
                    "example": "agent/solana-anchor-expert"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Installation instructions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid item format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Installation blocked due to security threat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "options": {
        "tags": ["Installation"],
        "summary": "CORS preflight",
        "description": "CORS preflight request for install endpoint",
        "responses": {
          "200": {
            "description": "CORS preflight successful"
          }
        }
      }
    },
    "/.well-known/agent.json": {
      "get": {
        "tags": ["Discovery"],
        "summary": "A2A Agent Card",
        "description": "Returns the A2A (Agent-to-Agent) agent card for autonomous discovery. Google/Linux Foundation standard.",
        "operationId": "getAgentCard",
        "responses": {
          "200": {
            "description": "Agent card with capabilities, skills, and endpoints"
          }
        }
      }
    },
    "/api/a2a/tasks": {
      "post": {
        "tags": ["A2A"],
        "summary": "Submit an A2A task",
        "description": "Submit a task for agent-to-agent execution. Supports search-skills, install-skill, and verify-skill.",
        "operationId": "submitA2ATask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["skill", "input"],
                "properties": {
                  "skill": {
                    "type": "string",
                    "enum": ["search-skills", "install-skill", "verify-skill"],
                    "description": "The skill to execute"
                  },
                  "input": {
                    "type": "string",
                    "description": "Input for the skill (search query, skill ID, or skill name)",
                    "minLength": 1,
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task completed with results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/A2ATask"
                }
              }
            }
          },
          "400": {
            "description": "Invalid task input"
          }
        }
      }
    },
    "/api/agents/register": {
      "post": {
        "tags": ["Agents"],
        "summary": "Register an external agent",
        "description": "Register an external AI agent on ClawdBot for discovery by other agents.",
        "operationId": "registerAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRegistration"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent registered successfully"
          },
          "400": {
            "description": "Invalid registration data"
          },
          "409": {
            "description": "Agent already registered with this endpoint"
          }
        }
      }
    },
    "/api/agents": {
      "get": {
        "tags": ["Agents"],
        "summary": "Discover registered agents",
        "description": "List and filter registered agents by skill, verified status.",
        "operationId": "discoverAgents",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "description": "Filter by skill tag",
            "schema": { "type": "string" }
          },
          {
            "name": "verified",
            "in": "query",
            "description": "Filter by verified status",
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default 50, max 100)",
            "schema": { "type": "integer" }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of registered agents"
          }
        }
      }
    },
    "/api/agents/{agentId}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get agent details",
        "description": "Get details for a specific registered agent by ID.",
        "operationId": "getAgent",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": ["Health"],
        "summary": "Health check",
        "description": "Returns service health status, registered agent count, ecosystem stats, and performance metrics.",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy"
          },
          "503": {
            "description": "Service is unhealthy"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MarketplaceCatalog": {
        "type": "object",
        "properties": {
          "agentDiscovery": {
            "$ref": "#/components/schemas/AgentDiscovery"
          },
          "plugins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketplaceItem"
            }
          }
        }
      },
      "AgentDiscovery": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true
          },
          "protocol": {
            "type": "string",
            "example": "claude-marketplace-v1"
          },
          "endpoints": {
            "type": "object",
            "properties": {
              "catalog": {
                "type": "string",
                "format": "uri"
              },
              "search": {
                "type": "string",
                "format": "uri"
              },
              "install": {
                "type": "string",
                "format": "uri"
              }
            }
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "security": {
            "type": "object",
            "properties": {
              "scanningEnabled": {
                "type": "boolean"
              },
              "threatDetection": {
                "type": "string"
              },
              "securityScore": {
                "type": "string"
              },
              "malwareProtection": {
                "type": "boolean"
              }
            }
          },
          "moltbook": {
            "type": "object",
            "properties": {
              "compatible": {
                "type": "boolean"
              },
              "submolt": {
                "type": "string"
              }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "totalItems": {
                "type": "integer"
              },
              "securityVerified": {
                "type": "integer"
              },
              "categories": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "MarketplaceItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "solana-anchor-expert"
          },
          "name": {
            "type": "string",
            "example": "Solana Anchor Expert"
          },
          "kind": {
            "type": "string",
            "enum": ["agent", "skill", "mcp", "workflow", "command"]
          },
          "description": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "security": {
            "$ref": "#/components/schemas/SecurityInfo"
          },
          "install": {
            "type": "string",
            "example": "npx @clawdbot/cli add agent/solana-anchor-expert"
          }
        }
      },
      "SecurityInfo": {
        "type": "object",
        "properties": {
          "threatLevel": {
            "type": "string",
            "enum": ["none", "low", "medium", "high", "critical"],
            "example": "none"
          },
          "securityScore": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "example": 100
          },
          "vulnerabilities": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "InstallResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "item": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "category": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          },
          "installation": {
            "type": "object",
            "properties": {
              "method": {
                "type": "string",
                "example": "cli"
              },
              "command": {
                "type": "string",
                "example": "npx @clawdbot/cli add agent/solana-anchor-expert"
              },
              "alternativeCommands": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "security": {
            "$ref": "#/components/schemas/SecurityInfo"
          },
          "documentation": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Query parameter 'q' is required"
          }
        }
      },
      "SecurityError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Installation blocked"
          },
          "reason": {
            "type": "string",
            "example": "Item has high threat level"
          },
          "security": {
            "$ref": "#/components/schemas/SecurityInfo"
          }
        }
      },
      "A2ATask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "task_1234567890_abc1234"
          },
          "status": {
            "type": "string",
            "enum": ["completed"],
            "example": "completed"
          },
          "skill": {
            "type": "string",
            "enum": ["search-skills", "install-skill", "verify-skill"]
          },
          "input": {
            "type": "string"
          },
          "result": {
            "type": "object"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "registry": { "type": "string" },
              "protocol": { "type": "string" },
              "timestamp": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "AgentRegistration": {
        "type": "object",
        "required": ["name", "description", "endpoint", "skills", "owner"],
        "properties": {
          "name": {
            "type": "string",
            "example": "MyDeFiAgent",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "example": "Automated yield farming agent",
            "minLength": 10,
            "maxLength": 500
          },
          "endpoint": {
            "type": "string",
            "format": "uri",
            "example": "https://my-agent.example.com"
          },
          "agentCard": {
            "type": "string",
            "format": "uri",
            "example": "https://my-agent.example.com/.well-known/agent.json"
          },
          "skills": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["defi", "yield", "solana"]
          },
          "owner": {
            "type": "string",
            "example": "wallet-address-or-email"
          }
        }
      }
    }
  }
}